lifeops_agent/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
sqlalchemy==2.0.31
openai==1.40.2
httpx>=0.27.0,<0.28.0
pypdf==4.2.0
pillow==10.4.0
pytesseract==0.3.10

